--- srml/support/procedural/tests/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "srml-support-procedural-tests"
-version = "0.1.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-
-[dependencies]
-hex-literal = { version = "0.1.0", optional = true }
-serde = { version = "1.0", default-features = false }
-serde_derive = { version = "1.0", optional = true }
-parity-codec = { version = "2.1", default-features = false }
-substrate-metadata = { path = "../../../metadata", default-features = false }
-sr-std = { path = "../../../../core/sr-std", default-features = false }
-sr-io = { path = "../../../../core/sr-io", default-features = false }
-sr-primitives = { path = "../../../../core/sr-primitives", default-features = false }
-srml-support-procedural = { path = ".." }
-srml-support = { path = "../.." }
-mashup = "0.1.7"
-pretty_assertions = "0.5.1"
-parity-codec-derive = { version = "2.1" }
-
-[features]
-default = ["std"]
-std = [
-	"hex-literal",
-	"serde/std",
-	"serde_derive",
-	"sr-io/std",
-	"parity-codec/std",
-	"sr-std/std",
-	"sr-primitives/std",
-	"substrate-metadata/std",
-]
-nightly = []
-strict = []
