+++ srml/support/procedural/tools/Cargo.toml
@@ -3,10 +3,8 @@ name = "srml-support-procedural-tools"
 version = "0.1.0"
 authors = ["Parity Technologies <admin@parity.io>"]
 
-[lib]
-proc-macro = true
-
 [dependencies]
+srml-support-procedural-tools-derive = { path = "./derive" }
 proc-macro2 = "0.4.24"
 quote = { version = "0.6.10", features = ["proc-macro"] }
 syn = { version = "0.15.21", features = ["proc-macro" ,"full", "extra-traits", "parsing"] }
